pyproject.toml
[project]
name = "social-media-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-alembic>=3.2.0",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.3.0",
    "flask-sqlalchemy-lite>=0.2.1",
    "marshmallow-sqlalchemy>=1.4.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
